price_producer/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
frozenlist==1.3.1
idna==3.4
kafka-python==2.0.2
multidict==6.0.2
yarl==1.8.1
